+++ cli/test/models/api_base_response.dart
@@ -1,5 +1,4 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
-import 'package:collection/collection.dart';
 
 part 'api_base_response.data.dart';