+++ cli/test/models/default_values.model.dart
@@ -1,5 +1,4 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
-import 'package:collection/collection.dart';
 
 part 'default_values.model.data.dart';