+++ cli/test/models/no_prefix_validation_test.model.dart
@@ -1,5 +1,4 @@
 import 'package:dataforge_annotation/dataforge_annotation.dart';
-import 'package:collection/collection.dart';
 
 part 'no_prefix_validation_test.model.data.dart';