+++ generator/example/lib/example.dart
@@ -208,8 +208,6 @@ class ComplexUser with _ComplexUser {
 
   factory ComplexUser.fromJson(Map<String, dynamic> json) =>
       _ComplexUser.fromJson(json);
-
-  
 }
 
 @Dataforge(deepCopyWith: true)